backend-django/requirements.txt
# Core Django dependencies
Django==5.0.1
django-ninja==1.1.0
anthropic==0.75.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
gunicorn==21.2.0
